requirements.txt
langchain==0.3.7
langgraph==0.2.45
accelerate==1.1.1
bitsandbytes==0.44.1
transformers==4.46.2
python-dotenv==1.0.1
python_telegram_bot==21.7
celery[redis]==5.4.0
